+++ R/calc_annual_stats.R
@@ -16,9 +16,9 @@
 #'    data set. Calculates statistics from all values, unless specified. Returns a tibble with statistics.
 #'
 #' @param data Data frame of daily data that contains columns of dates, flow values, and (optional) groups (e.g. station numbers).
-#'    Leave blank if using \code{station_number} argument.
+#'    Leave blank or set to \code{NULL} if using \code{station_number} argument.
 #' @param dates Name of column in \code{data} that contains dates formatted YYYY-MM-DD. Only required if dates column name is not 
-#'    'Date' (default). Leave blank if using \code{station_number} argument.
+#'    'Date' (default). Leave blank or set to \code{NULL} if using \code{station_number} argument.
 #' @param values Name of column in \code{data} that contains numeric flow values, in units of cubic metres per second.
 #'    Only required if values column name is not 'Value' (default). Leave blank if using \code{station_number} argument.
 #' @param groups Name of column in \code{data} that contains unique identifiers for different data sets, if applicable. Only required 
@@ -35,11 +35,14 @@
 #' @param percentiles Numeric vector of percentiles to calculate. Set to \code{NA} if none required. Default \code{c(10,90)}.
 #' @param water_year_start Numeric value indicating the month (\code{1} through \code{12}) of the start of water year for
 #'    analysis. Default \code{1}.
-#' @param start_year Numeric value of the first year to consider for analysis. Leave blank to use the first year of the source data.
-#' @param end_year Numeric value of the last year to consider for analysis. Leave blank to use the last year of the source data.
-#' @param exclude_years Numeric vector of years to exclude from analysis. Leave blank to include all years.             
-#' @param months Numeric vector of months to include in analysis (e.g. \code{6:8} for Jun-Aug). Leave blank to summarize 
-#'    all months (default \code{1:12}).
+#' @param start_year Numeric value of the first year to consider for analysis. Leave blank or set well before start date (i.e.
+#'    \code{1800}) to use from the first year of the source data.
+#' @param end_year Numeric value of the last year to consider for analysis. Leave blank or set well after end date (i.e.
+#'    \code{2100}) to use up to the last year of the source data.
+#' @param exclude_years Numeric vector of years to exclude from analysis. Leave blank or set to \code{NULL} to include all years.             
+#' @param months Numeric vector of months to include in analysis. For example, \code{3} for March, \code{6:8} for Jun-Aug or 
+#'    \code{c(10:12,1)} for first four months (Oct-Jan) when \code{water_year_start = 10} (Oct). Default summarizes all 
+#'    months (\code{1:12}).
 #' @param transpose Logical value indicating whether to transpose rows and columns of results. Default \code{FALSE}.
 #' @param ignore_missing Logical value indicating whether dates with missing values should be included in the calculation. If
 #'    \code{TRUE} then a statistic will be calculated regardless of missing dates. If \code{FALSE} then only those statistics from