+++ R/plot_daily_cumulative_stats.R
@@ -66,6 +66,7 @@ plot_daily_cumulative_stats <- function(data,
                                         exclude_years, 
                                         months = 1:12,
                                         log_discharge = FALSE,
+                                        log_ticks = ifelse(log_discharge, TRUE, FALSE),
                                         include_title = FALSE,
                                         add_year){
   
@@ -95,6 +96,7 @@ plot_daily_cumulative_stats <- function(data,
   }
   
   log_discharge_checks(log_discharge) 
+  log_ticks_checks(log_ticks, log_discharge)
   add_year_checks(add_year)
   include_title_checks(include_title)  
   
@@ -216,7 +218,7 @@ plot_daily_cumulative_stats <- function(data,
              ggplot2::scale_color_manual(values = c("Median" = "purple3", "Mean" = "springgreen4")) +
              {if (!log_discharge) ggplot2::scale_y_continuous(expand = c(0, 0), breaks = scales::pretty_breaks(n = 7))} +
              {if (log_discharge) ggplot2::scale_y_log10(expand = c(0, 0), breaks = scales::log_breaks(n = 8, base = 10) )} +
-             {if (log_discharge) ggplot2::annotation_logticks(base= 10, "left", colour = "grey25", size = 0.3,
+             {if (log_discharge & log_ticks) ggplot2::annotation_logticks(base= 10, "left", colour = "grey25", size = 0.3,
                                                               short = ggplot2::unit(.07, "cm"), mid = ggplot2::unit(.15, "cm"),
                                                               long = ggplot2::unit(.2, "cm"))} +
              ggplot2::scale_x_date(date_labels = "%b", date_breaks = "1 month",