+++ R/plot_longterm_daily_stats.R
@@ -20,7 +20,8 @@
 #'
 #' @inheritParams calc_longterm_daily_stats
 #' @inheritParams plot_annual_stats
-#' @param add_year Numeric value indicating a year of daily flows to add to the daily statistics plot. Leave blank for no years.
+#' @param add_year Numeric value indicating a year of daily flows to add to the daily statistics plot. Leave blank
+#'    or set to \code{NULL} for no years.
 #' @param include_extremes Logical value to indicate plotting a ribbon with the range of daily minimum and maximum flows. 
 #'    Default \code{TRUE}.
 #' @param inner_percentiles Numeric vector of two percentile values indicating the lower and upper limits of the 
@@ -77,6 +78,7 @@ plot_longterm_daily_stats <- function(data,
                                       outer_percentiles = c(5,95),
                                       add_year,
                                       log_discharge = TRUE,
+                                      log_ticks = ifelse(log_discharge, TRUE, FALSE),
                                       include_title = FALSE){
   
   ## ARGUMENT CHECKS
@@ -102,6 +104,7 @@ plot_longterm_daily_stats <- function(data,
   }
   
   log_discharge_checks(log_discharge)
+  log_ticks_checks(log_ticks, log_discharge)
   include_title_checks(include_title)  
   ptile_ribbons_checks(inner_percentiles, outer_percentiles)
   
@@ -262,7 +265,7 @@ plot_longterm_daily_stats <- function(data,
         ggplot2::geom_point(ggplot2::aes(y = Median), size = 2, na.rm = TRUE, colour = "dodgerblue4") +
         {if(!log_discharge) ggplot2::scale_y_continuous(expand = c(0, 0), breaks = scales::pretty_breaks(n = 8))}+
         {if(log_discharge) ggplot2::scale_y_log10(expand = c(0, 0), breaks = scales::log_breaks(n = 8, base = 10))} +
-        {if(log_discharge) ggplot2::annotation_logticks(base = 10, "l", colour = "grey25", size = 0.3, short = ggplot2::unit(0.07, "cm"),
+        {if(log_discharge & log_ticks) ggplot2::annotation_logticks(base = 10, "l", colour = "grey25", size = 0.3, short = ggplot2::unit(0.07, "cm"),
                                                         mid = ggplot2::unit(0.15, "cm"), long = ggplot2::unit(0.2, "cm"))} +
         ggplot2::scale_x_discrete(expand = c(0.01,0.01)) +
         ggplot2::ylab(y_axis_title) +